@@ -35,7 +35,7 @@ class LimeLightCRMTest extends TestCase
      */
     protected $streamMock;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
