@@ -1,4 +1,5 @@
 <?php
+
 namespace KevinEm\LimeLightCRM\Tests\v1;
 
 use GuzzleHttp\Client;
@@ -6,41 +7,41 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 use KevinEm\LimeLightCRM\Exceptions\LimeLightCRMGenericException;
+use KevinEm\LimeLightCRM\Exceptions\LimeLightCRMParseResponseException;
 use KevinEm\LimeLightCRM\v1\LimeLightCRM;
 use PHPUnit\Framework\TestCase;
 
 class LimeLightCRMTest extends TestCase
 {
-    /**
-     * @expectedException \KevinEm\LimeLightCRM\Exceptions\LimeLightCRMParseResponseException
-     * @expectedExceptionMessage Cannot understand limelight response
-     */
     public function testExceptionOnUnparseableResponse()
     {
-        $rawExpectedResponse  = 'alkjsdfklasdjlkf';
-        $mock = new MockHandler([
+        $this->expectException(LimeLightCRMParseResponseException::class);
+        $this->expectExceptionMessage('Cannot understand limelight response');
+
+        $rawExpectedResponse = 'alkjsdfklasdjlkf';
+        $mock                = new MockHandler([
             new Response(200, [], $rawExpectedResponse),
         ]);
 
         $handler = HandlerStack::create($mock);
-        $client = new Client(['handler' => $handler]);
+        $client  = new Client(['handler' => $handler]);
 
-        $s = new LimeLightCRM($client, []);
+        $s = new LimeLightCRM($client, ['base_url' => '', 'username' => '', 'password' => '']);
 
         $s->makeRequest('', [], 'POST');
     }
 
     public function testResponseObjectReturns()
     {
-        $rawExpectedResponse  = '{"response_code": "100","error_found": "0"}';
-        $mock = new MockHandler([
+        $rawExpectedResponse = '{"response_code": "100","error_found": "0"}';
+        $mock                = new MockHandler([
             new Response(200, [], $rawExpectedResponse),
         ]);
 
         $handler = HandlerStack::create($mock);
-        $client = new Client(['handler' => $handler]);
+        $client  = new Client(['handler' => $handler]);
 
-        $s = new LimeLightCRM($client, []);
+        $s = new LimeLightCRM($client, ['base_url' => '', 'username' => '', 'password' => '']);
 
         $res = $s->makeRequest('', [], 'POST');
 
@@ -50,15 +51,15 @@ class LimeLightCRMTest extends TestCase
 
     public function testExceptionOnErrorCode()
     {
-        $rawExpectedResponse  = '{"response_code": "200","error_found": "Something went wrong"}';
-        $mock = new MockHandler([
+        $rawExpectedResponse = '{"response_code": "200","error_found": "Something went wrong"}';
+        $mock                = new MockHandler([
             new Response(200, [], $rawExpectedResponse),
         ]);
 
         $handler = HandlerStack::create($mock);
-        $client = new Client(['handler' => $handler]);
+        $client  = new Client(['handler' => $handler]);
 
-        $s = new LimeLightCRM($client, []);
+        $s = new LimeLightCRM($client, ['base_url' => '', 'username' => '', 'password' => '']);
 
         try {
             $s->makeRequest('', [], 'POST');
@@ -73,15 +74,15 @@ class LimeLightCRMTest extends TestCase
 
     public function test343IsSuccess()
     {
-        $rawExpectedResponse  = '{"response_code": "343","error_found": "Everything is fine"}';
-        $mock = new MockHandler([
+        $rawExpectedResponse = '{"response_code": "343","error_found": "Everything is fine"}';
+        $mock                = new MockHandler([
             new Response(200, [], $rawExpectedResponse),
         ]);
 
         $handler = HandlerStack::create($mock);
-        $client = new Client(['handler' => $handler]);
+        $client  = new Client(['handler' => $handler]);
 
-        $s = new LimeLightCRM($client, []);
+        $s = new LimeLightCRM($client, ['base_url' => '', 'username' => '', 'password' => '']);
 
         $resp = $s->makeRequest('', [], 'POST');
 
