@@ -1,10 +1,12 @@
 <?php
+
 namespace KevinEm\LimeLightCRM\Tests\v2;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
+use KevinEm\LimeLightCRM\Exceptions\LimeLightCRMGenericException;
 use KevinEm\LimeLightCRM\v2\LimeLightCRM;
 use PHPUnit\Framework\TestCase;
 
@@ -12,15 +14,15 @@ class LimeLightCRMTest extends TestCase
 {
     public function testRequestSuccess()
     {
-        $rawExpectedResponse  = '{"status": "SUCCESS"}';
-        $mock = new MockHandler([
+        $rawExpectedResponse = '{"status": "SUCCESS"}';
+        $mock                = new MockHandler([
             new Response(200, [], $rawExpectedResponse),
         ]);
 
         $handler = HandlerStack::create($mock);
-        $client = new Client(['handler' => $handler]);
+        $client  = new Client(['handler' => $handler]);
 
-        $s = new LimeLightCRM($client, []);
+        $s = new LimeLightCRM($client, ['base_url' => '', 'username' => '', 'password' => '']);
 
         $response = $s->makeRequest('', []);
 
@@ -29,22 +31,21 @@ class LimeLightCRMTest extends TestCase
         $this->assertEquals($jsonExpectedResponse, $response);
     }
 
-    /**
-     * @expectedException \KevinEm\LimeLightCRM\Exceptions\LimeLightCRMGenericException
-     * @expectedExceptionMessage error message
-     * @expectedExceptionCode 400
-     */
     public function testRequestErrorWithMessage()
     {
-        $rawExpectedResponse  = '{"status": "SUCCESS","message": "error message"}';
-        $mock = new MockHandler([
+        $this->expectException(LimeLightCRMGenericException::class);
+        $this->expectExceptionMessage('error message');
+        $this->expectExceptionCode(400);
+
+        $rawExpectedResponse = '{"status": "SUCCESS","message": "error message"}';
+        $mock                = new MockHandler([
             new Response(400, [], $rawExpectedResponse),
         ]);
 
         $handler = HandlerStack::create($mock);
-        $client = new Client(['handler' => $handler]);
+        $client  = new Client(['handler' => $handler]);
 
-        $s = new LimeLightCRM($client, []);
+        $s = new LimeLightCRM($client, ['base_url' => '', 'username' => '', 'password' => '']);
 
         $s->makeRequest('', []);
     }
